Backend/requirements.txt
fastapi
uvicorn
requests
python-dotenv
pydantic
newspaper3k
lxml
lxml_html_clean
supabase
twelvedata
pandas
sqlite-utils
pytz
